requirements.txt
autogluon.tabular==1.5.0
pandas>=2.0
numpy<2
scikit-learn>=1.3
torch>=2.0
lightgbm
ray>=2.43.0,<2.53.0